+++ frontend/src/pages/ReviewWritingPage/layout/components/QnABoxLayout/style.ts
